New API: a single options object
